disc-ripper: CDDB.pm now lazy loaded and optional as such..
Yet CDDB.pm was a prerequiste and the ripper would have failed
if loading would have failed.
This is completely stupid behaviour given that we have all sorts
of fake entry fallback stuff at hand anyway!

So lazy load it instead if needed, and simply don't fail if
it's not available.