backup.pl: BUGFIX: fix git(1) stash backups..
Found out in a somewhat messy way that git(1) stash backups as
stored in the repo (and thus as present in bundles) are rather
useless by themselves, because of the missing file
.git/logs/refs/stash, which is necessary for this fragile shell
script to work properly.

So backup.pl now also takes care of this one and manages it
as BUNDLENAME.stashlog as necessary.