backup.pl: FIX/tweak for git(1) stash backups..
The commit (backup.pl: BUGFIX: fix git(1) stash backups..,
2012-02-16) also saves the stashlog so that stash objects in
bundles can be somewhat easily used.

However, it may be that a backup first creates a bundle which
contains a stash, and the next backup may create a bundle
without a stash, maybe because that has been evolved to a real
commit.
In such a case the .stashlog file yet wasn't deleted even though
it refers to a non-existent bundle.
Fix that.