backup.pl: rewrite Archive::create() for total tar(1) compat..
While playing around some more with tar implementations i found
out that the current approach for the complete backup, i.e.,
writing all paths that should be stored into the archive and
then using this file as input to tar(1), is not compatible at
all.

So instead of fooling around some more with command line options
to gain at least compatibility with some current tar(1)s
i decided to rewrite Archive::create() in a completely neutral
way, meaning to use 'xargs | tar -r' and using an explicit bzip2
on the resulting archive afterwards in --complete case.
This solution is still not optimal, though.