gen-uushar-wrapper: added - enwrap exec environment..
This is a nice little script which generates an encapsulating
execution environment.
It creates a shell script that contains an uuencode(1)d shell archive,
and is capable to run the programs therein.

I use it to have several versions of sccs, of which none is
really installed, for testing purposes.