gen-uushar-wrapper: ISC license, ARGV enable..
Rethought it a bit and it turned out to be stupid.
First of all, let's use the ISC license, as i'm currently
turning over all my work to it.

Then enable the possibility to specify the commands which should
make up the archive on the command line, thus allow
tab-completion.

Also pimp the output that is produced and group it a bit in
blocks.

Introduce the produced file with a TOC of all programs which are
contained, just like (uu)shar(1) does.

And finally, get rid of the functions, unroll all the code in
the archive file, since it's only a few lines anyway.
While at it, drop support for real execution of a command the
first time an archive is executed, since sh(1) doesn't support
goto:, and there is no more function to call.

The good news on the latter: well, arguments should now be
passed through to the actual command unchanged and thus as
desired.  Au.