backup.pl: git-update-index before check if sth. to stash..
Yet the index was not refreshed before diff-index, diff-files
and ls-files -o -z ran to check wether a stash has to be produced.
This caused a stash to be produced though no changes were to
stash away, so that the final stash pop failed due to a lack
of something to pop.
Since the code reflects one-on-one what git-stash.sh does,
i guess the update-index should be applied before someone
checks via diff-index.  Logical.