gen-uushar-wrapper: add compression and [^X#] egrep chance..
Add optional --(compress|gzip|bzip2|xz) arguments that will
enforce compression of the shell archive members.

And because the generated script is still a shell archive,
add the replacement of ^ with X via sed(1), so that it is
possible to search all executed lines of the script via
"$ egrep -v ^[X#] shell-archive", as that is what people
are doing for decades to look into shell archives.

While at it, remove some tabulators and tweak messages.