gen-uushar-wrapper: move around atexit trap plus..
set -C is used to avoid overwriting by redirection,
but the atexit trap radically removes $ARCH.
Thus, whereas the script would have failed to overwrite it,
it forcefully would have removed it instead.
(It is not that much an issue since we bail out long before
if the file exists, but an atomicity problem remains.)

So move the atexit trap installation to after the first
write to $ARCH, once it is clear it belongs to us.
While here, replace an if with an AND-OR list.