disc-ripper: add utf8ify(), use PerlIO encoding layer..
Yeah, finally get rid of this hacky Encode::_utf8_off()
crap that is spreaded around everywhere, but do a call-in
to utf8ify() and use PerlIO layer for the text files.
Like this CDDB can be as shitty as it want, we get it right,
and from now on in a clean way.