s-postgray: diversify SIGHUP action; radical tighten FreeBSD sandbox..
We now make the actual action of SIGHUP _optionally_ dependent
upon --untamed mode (if VAL_OS_SANDBOX>0, of course).

Use this possibility on FreeBSD, and instead of having a FD to
/ open, and using realpath(3) beforehand to have all the paths for
later openat(2), only keep a FD to --store-path open, and only
support file operations within this effective "chroot(2)"!
Allows us to get rid of any path handling on FreeBSD.
On a FreeBSD ML thread i said

  20230413192625.mUQ_T%steffen@sdaoden.eu

  I mean, i can evaluate configuration in a/the "super-capable" base
  process, and then simply fork off a new server which then inherits
  the new configuration (after the old has been told to die), but
  that is a real mess.
  Also because, you know, so i opened a directory FD for / (the way
  i do it: do this, use realpath(3) on all paths, and then simply
  openat(2) "rootfd,&[1]" to not openat(2) an absolute path..), but
  this is only for the sandboxed process.

but when waking up this morning i told me that having the super
tight capsicum(4) cage is a value by itself.  And you can always
  pkill -TERM postgray
or use --shutdown + --startup if you will.