s-dkim-sign: -# must be first argument..
It can happen that -# runs nicely when run by root, but that the
real usage fails because the anon-user which execs has not the
necessary rights to access the configuration file.
In that case the error message would have caused faulty interpretation
as milter communication, causing in the program being killed.

We can hardly automatize the solution for this.  Instead, require
that --test-mode must be the first argument, and immediately open
the syslog connection and henceforth log there, then